srcs/backend/transa/pyproject.toml
[project]
name = "transa"
version = "0.1.0"
description = "Backend part for ft_transcendence"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels==4.2.0",
    "channels_redis>=4.1.0",
    "daphne==4.1.2",
    "django-filter==25.1",
    "django==5.1.6",
    "djangorestframework==3.15.2",
    "gunicorn==22.0.0",
    "markdown==3.7",
    "psycopg2-binary==2.9.10",
    "pillow==11.1.0",
    "djangorestframework-simplejwt>=5.5.0",
    "requests>=2.32.3",
    "pyotp>=2.9.0",
    "qrcode>=8.0",
    "django-cors-headers>=4.7.0",
    "redis>=5.2.1",
    "django-stubs>=5.2.0",
    "websockets>=12.0",
    "django-extensions>=4.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.9.7",
]
